fix(bindings): align struct layouts with Box2D v3.0.0
Correct BodyDef offsets and add MotionLocks.

Correct ShapeDef offsets and add EnableCustomFiltering.

Update SurfaceMaterial.UserMaterialId to ulong.

Resolves NaN physics positions.

+++ src/Box2DBindings/SurfaceMaterial.cs
@@ -49,7 +49,7 @@ public struct SurfaceMaterial
     /// User material identifier. This is passed with query results and to friction and restitution
     /// combining functions. It is not used internally.
     /// </summary>
-    public int UserMaterialId;
+    public ulong UserMaterialId;
 
     /// <summary>
     /// Custom debug draw color.
